backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
flake8-isort==6.0.0
Pillow==9.0.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
reportlab==4.2.0
gunicorn==20.1.0